Update 1.8
Added new functionality's

App.java: Added a "\n" at the end of the error message to have it better readable in the text file

TimerApp.java: Added a System.out.println to be shown in the terminal to point out the program is running.

TimerJob.java:

Terminal command line input has been archieved (Has to be perfected(At the moment it requires the value's to be given in two lines)

It can read the FOREX Codes from a file that have been seperated by a new line.

It can not yet recognise the FOREX Codes when they are together and it keeps printing the error message from App.java (To be fixed ASAP)