Update 1.9
Added new functionality's

TimerApp.java: Added a System.out.println to be shown in the terminal to point out the program is waiting for input if any is necessary.

TimerJob.java:

It can read the FOREX Codes from a file that have been are not separated by a new line.